build out day 0 multi coin branch
